move custom http client creation with env proxy args into same package as other network related helpers
