+++ builder/qemu/step_set_iso.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"net/http"
 
-	commonhelper "github.com/hashicorp/packer/helper/common"
+	"github.com/hashicorp/packer/common/net"
 	"github.com/hashicorp/packer/helper/multistep"
 	"github.com/hashicorp/packer/packer"
 )
@@ -29,7 +29,7 @@ func (s *stepSetISO) Run(ctx context.Context, state multistep.StateBag) multiste
 
 		req.Header.Set("User-Agent", "Packer")
 
-		httpClient := commonhelper.HttpClientWithEnvironmentProxy()
+		httpClient := net.HttpClientWithEnvironmentProxy()
 
 		res, err := httpClient.Do(req)
 		if err == nil && (res.StatusCode >= 200 && res.StatusCode < 300) {
